+++ spec/unit/config_spec.js
@@ -150,6 +150,36 @@ describe('Config', function () {
     })
   );
 
+  describe('keepAlive', () => {
+    it('defaults to disabled', function () {
+      let config = new Config({
+        merchantId: 'merchantId',
+        publicKey: 'publicKey',
+        privateKey: 'privateKey',
+        environment: 'development'
+      });
+
+      assert.equal(config.keepAliveAgentEnabled, false);
+      assert.equal(config.keepAliveConfig.keepAlive, true);
+    });
+
+    it('is enabled when user passes in a config', function () {
+      let config = new Config({
+        merchantId: 'merchantId',
+        publicKey: 'publicKey',
+        privateKey: 'privateKey',
+        environment: 'development',
+        keepAlive: {
+          keepAliveMsecs: 9001
+        }
+      });
+
+      assert.equal(config.keepAliveAgentEnabled, true);
+      assert.equal(config.keepAliveConfig.keepAlive, true);
+      assert.equal(config.keepAliveConfig.keepAliveMsecs, 9001);
+    });
+  });
+
   describe('accessToken', function () {
     it('uses accessToken parsed environment', function () {
       let config = new Config({