+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+* Add keep-alive to http module
+
 ## 2.22.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors
 * Add GraphQL ID to `CreditCardVerification`, `Customer`, `Dispute`, and `Transaction`